+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 
 - Updated to 1.21
 - Removed the config for limiting based on NBT size
+- Removed the config for showing NBT in tooltip
 - Changed the config for max location size to be unlimited by default